+++ odoo/songs/updates/update_10_1_2/update.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 # Copyright 2017 Camptocamp SA
-# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html)
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 
 import anthem
 
@@ -29,7 +29,45 @@ def update_crm_stage_values(ctx):
 
 
 @anthem.log
+def update_vouchers(ctx):
+
+    vouchers = ctx.env['sale.discount.program'].search([
+        ('voucher_code', '!=', False)])
+    gift = vouchers.filtered(lambda v: not v.partner_id)
+    vouchers -= gift
+    gift.write({
+        'type': 'gift_voucher'
+    })
+    vouchers.write({
+        'type': 'voucher'
+    })
+
+
+@anthem.log
+def update_promo_codes(ctx):
+
+    promo_codes = ctx.env['sale.discount.program'].search([
+        ('promo_code', '!=', False)])
+    promo_codes.write({
+        'type': 'promo_code'
+    })
+
+
+@anthem.log
+def update_discount_programs(ctx):
+
+    programs = ctx.env['sale.discount.program'].search([
+        ('automatic', '=', True)])
+    programs.write({
+        'type': 'discount_program'
+    })
+
+
+@anthem.log
 def main(ctx):
     """ Main: update 10.1.2 """
     clean_groups(ctx)
     update_crm_stage_values(ctx)
+    update_discount_programs(ctx)
+    update_promo_codes(ctx)
+    update_vouchers(ctx)