+++ odoo/songs/updates/update_10_1_2/update.sql
@@ -14,3 +14,12 @@ WHERE
     module = 'specific_payment_mode'
 AND lang = 'fr_FR'
 AND value LIKE 'L''apport doit être%';
+
+-- Update XMLID from discount_program_report module data
+
+UPDATE ir_model_data
+SET name = 'voucher_config_template_fr',
+    noupdate = true
+WHERE module = 'discount_program_report'
+AND name = 'config_template_fr'
+AND model = 'sale.discount.program.report.config';