+++ HISTORY.rst
@@ -11,6 +11,7 @@ Unreleased
 * DT-71: Simplify usage of depiltech payment mode
 * DT-75: Custom displaying of pages on partner form view
 * DT-73 / DT-77: Custom lead workflow than lead -> opportunity -> customer
+* Replace gift_voucher boolean by program_type selection
 
 **Bugfixes**